Create first struct
Added color.rb to structs as method to port the raylib-c color struct to
ruby.

It is the first attempt to make a port of a struct and will serve to
show how to do so in the future presuming bugs arise due to potential
improper implementation as shown when the basic color constants are
established.