Add color initialization method
Added struct field accessors and initialization method for ease of use.

Need to quickly intialize a color with all the values at once rather
than intialize each field by itself.

May be a bit hacky as super is not called; super causes errors and
depcreation message as FFI::Struct uses it for layout whilst we use it
for intializing fields.

Alternative is a new_<struct_name> method. E.g. new_color. The debuff of
this is that it is non-standardized.