Write basic window example
This example is a port of the general first example of raylib-c.

Whilst not a currently functioning example it provides a blue print of
how the functions of raylib-ruby should be able to be called in use.
Otherwise it provides, as intended, a basic example to test whether
raylib is currently working.